hscripted state changes
Changed the class name to "HScriptedState" for consistency and because it better describes its purpose.
The constructor now takes a script instead of a string path key.
So to create a scripted state instance there's now the fromString and fromFile functions.
Also removed the additionalVars argument because I feel like, in practice, it either won't get used or it'll be the same to just set the variables directly. Not too sure abt it tho!